Move repair_after_upgrade_test to upgrade_tests package
